Added period flushing of labels to disk.
