+++ snorkel/contrib/snark/annotations.py
@@ -1,46 +1,10 @@
 from functools import partial
-from pickle import loads
 
-from models.candidate import Candidate, wrap_candidate
-from models.context import Sentence, Span
+from models.candidate import wrap_candidate
 from snorkel.annotations import load_label_matrix
 from snorkel.models.annotation import Label, LabelKey
 from snorkel.models.meta import snorkel_conn_string
 from snorkel.models.views import create_serialized_candidate_view
-
-
-class SparkCorpusParser:
-    """
-    Distributes raw documents to a Spark cluster and applies a parser to them,
-    returning a hierarchy of Context objects. See snorkel.parser.CorpusParser.
-
-    NOTE: Currently just a stub.
-    """
-    def __init__(self, snorkel_session, spark_session, parser, fn=None):
-        """
-        Constructor
-
-        :param snorkel_session: the SnorkelSession for the Snorkel application
-        :param spark_session: a PySpark SparkSession
-        :param parser: See snorkel.parser.Parser
-        :param fn: Function to apply to parser output
-        """
-        self.snorkel_session = snorkel_session
-        self.spark_session = spark_session
-        self.parser = parser
-        self.fn = fn
-
-    def apply(self, docs, **kwargs):
-        # TODO: Need to be able to start a Parser server on each server...
-        pass
-
-    def _clear(self):
-        # TODO
-        # session.query(Context).delete()
-        # # We cannot cascade up from child contexts to parent Candidates,
-        # # so we delete all Candidates too
-        # session.query(Candidate).delete()
-        pass
 
 
 class SparkLabelAnnotator:
@@ -100,7 +64,12 @@ class SparkLabelAnnotator:
                 label_tuples.append({
                     'candidate_id': cid, 'key_id': lf_id, 'value': value})
 
-        # Bulk insert the labels
+                # Periodically flushes the labels to disk
+                if len(label_tuples) >= 100000:
+                    self.snorkel_session.execute(Label.__table__.insert(), label_tuples)
+                    label_tuples = []
+
+        # Flushes the remaining labels
         self.snorkel_session.execute(Label.__table__.insert(), label_tuples)
 
         # Return label matrix from the Snorkel DB
